Refactor Python preprocessing into modules, use bibtexparser, fix CI
Split the monolithic __init__.py into venues, authors, bibtex,
publications, and cli modules. Replace hand-rolled BibTeX parser with
bibtexparser v2. Add hugo-build CLI entrypoint for production builds.
Fix GitHub Actions workflow to use uv + hugo-build so publications.json
is generated in CI.
